+++ df-chat-enhance/src/archive/DFChatArchive.ts
@@ -132,8 +132,10 @@ export class DFChatArchive {
 	private static async _generateChatArchiveFile(id: number, name: string, chats: ChatMessage[] | ChatMessageData[], visible: boolean): Promise<DFChatArchiveEntry> {
 		// Get the folder path
 		const folderPath = SETTINGS.get<string>(this.PREF_FOLDER);
+		// Replace special characters in name to underscores
+		const safeName = name.replace(/[^ a-z0-9-_()[\]<>]/gi, '_');
 		// Generate the system safe filename
-		const fileName = encodeURI(`${id}_${name}.json`);
+		const fileName = encodeURI(`${id}_${safeName}.json`);
 		// Create the File and contents
 		const file = new File([JSON.stringify(chats, null, '')], fileName, { type: 'application/json' });
 		const response: { path?: string; message?: string } = <any>await FilePicker.upload(this.DATA_FOLDER, folderPath, file);