+++ df-chat-enhance/CHANGELOG.md
@@ -1,5 +1,8 @@
 # DF Chat Enhancements
 
+# Release 3.7.0 (2022-05-30)
+- **NEW #295:** Send message button has been added to the chat box. This can be enabled in the module settings.
+
 # Release 3.6.1 (2022-05-30)
 - **FIX #231:** Conflict with the combat round messages posted by Monk's Little Details.
 - **FIX #370:** Chat Archive will now handle special characters in archive names.